Se cambio modelo de datos
Solo compila pero no funciona aun

+++ Flor Shop/Model/coredata/CartDetail.swift
@@ -9,7 +9,7 @@ import Foundation
 
 struct CartDetail: Identifiable {
     let id: UUID
-    let quantity: Double
+    let quantity: Int64
     let subtotal: Double
     let product: Product
 }
